cron_scheduler/requirements.txt
psycopg2-binary
python-dotenv
boto3